ISSUE#3 refactor: input component에 label까지  받을 수 있도록 리팩토링
