--- src/Components/Atoms/Input/Input.tsx
@@ -1,9 +1,0 @@
-import { InputHTMLAttributes } from "react";
-
-import { StyledInput } from "./Input.style";
-
-const Input = (inputProps: InputHTMLAttributes<HTMLInputElement>) => {
-	return <StyledInput {...inputProps} />;
-};
-
-export default Input;
