+++ src/Components/Molecules/LoginForm/LoginForm.tsx
@@ -1,7 +1,7 @@
 import { FormEvent } from "react";
 
 import { LoginButton, StyledForm } from "./LoginForm.style";
-import Input from "../../Atoms/Input/Input";
+import Input from "../Input/Input";
 
 const LoginForm = () => {
 	const handleSubmit = (e: FormEvent<HTMLFormElement>) => {
@@ -11,8 +11,8 @@ const LoginForm = () => {
 
 	return (
 		<StyledForm onSubmit={handleSubmit}>
-			<Input placeholder="아이디" />
-			<Input placeholder="비밀번호" type="password" />
+			<Input inputProps={{ placeholder: "이메일" }} />
+			<Input inputProps={{ placeholder: "비밀번호", type: "password" }} />
 			<LoginButton>로그인</LoginButton>
 		</StyledForm>
 	);
